feat: enhance build
